test: add getPermitSignature for hardhat test
