+++ test/helpers/utils/getPermitSignature.ts
@@ -1,0 +1,131 @@
+import { BigNumberish, Signature } from "ethers";
+import { splitSignature } from "ethers/lib/utils";
+import { CoinBridgeToken, USDCRealT } from "../../../typechain";
+
+export async function getPermitSignatureRealToken(
+  // wallet: Wallet,
+  wallet: any,
+  spender: string,
+  value: BigNumberish,
+  deadline: BigNumberish,
+  token: CoinBridgeToken | USDCRealT,
+  permitConfig?: {
+    nonce?: BigNumberish;
+    name?: string;
+    chainId?: number;
+    version?: string;
+  }
+): Promise<Signature> {
+  const [nonce, name, version, chainId] = await Promise.all([
+    permitConfig?.nonce ?? token.nonces(wallet.address),
+    permitConfig?.name ?? token.name(),
+    permitConfig?.version ?? "2",
+    permitConfig?.chainId ?? wallet.getChainId(),
+  ]);
+
+  return splitSignature(
+    await wallet._signTypedData(
+      {
+        name,
+        version,
+        chainId,
+        verifyingContract: token.address,
+      },
+      {
+        Permit: [
+          {
+            name: "owner",
+            type: "address",
+          },
+          {
+            name: "spender",
+            type: "address",
+          },
+          {
+            name: "value",
+            type: "uint256",
+          },
+          {
+            name: "nonce",
+            type: "uint256",
+          },
+          {
+            name: "deadline",
+            type: "uint256",
+          },
+        ],
+      },
+      {
+        owner: wallet.address,
+        spender,
+        value,
+        nonce,
+        deadline,
+      }
+    )
+  );
+}
+
+export async function getPermitSignatureERC20(
+  // wallet: Wallet,
+  wallet: any,
+  spender: string,
+  value: BigNumberish,
+  deadline: BigNumberish,
+  token: CoinBridgeToken | USDCRealT,
+  permitConfig?: {
+    nonce?: BigNumberish;
+    name?: string;
+    chainId?: number;
+    version?: string;
+  }
+): Promise<Signature> {
+  const [nonce, name, version, chainId] = await Promise.all([
+    permitConfig?.nonce ?? token.nonces(wallet.address),
+    permitConfig?.name ?? token.name(),
+    permitConfig?.version ?? "1",
+    permitConfig?.chainId ?? wallet.getChainId(),
+  ]);
+
+  return splitSignature(
+    await wallet._signTypedData(
+      {
+        name,
+        version,
+        chainId,
+        verifyingContract: token.address,
+      },
+      {
+        Permit: [
+          {
+            name: "owner",
+            type: "address",
+          },
+          {
+            name: "spender",
+            type: "address",
+          },
+          {
+            name: "value",
+            type: "uint256",
+          },
+          {
+            name: "nonce",
+            type: "uint256",
+          },
+          {
+            name: "deadline",
+            type: "uint256",
+          },
+        ],
+      },
+      {
+        owner: wallet.address,
+        spender,
+        value,
+        nonce,
+        deadline,
+      }
+    )
+  );
+}
